fix: resolves conflicts arising from the merge between financial analysis and CRUD (Credit Rights Application Development) of transactions.
- Fixes FinancialAnalysisService (search by period, removes dead code)
- Unifies TransactionClassificationService into TransactionService
- Fixes data type and integration with the AI ​​mock
- Removes duplicate endpoint /classify-transactions

+++ backend/src/main/java/br/com/financeai/entity/FinancialAnalysis.java
@@ -39,7 +39,7 @@ public class FinancialAnalysis {
     private Long id;
 
     @Column(nullable = false)
-    private Integer nivelEndividamento;
+    private BigDecimal nivelEndividamento;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
