+++ backend/src/main/java/br/com/financeai/integration/dto/response/MlResponse.java
@@ -12,7 +12,7 @@ public record MlResponse(
 
         FinancialProfile financialProfile,
 
-        Integer nivelEndividamento,
+        BigDecimal nivelEndividamento,
 
         SavingFrequency frequenciaPoupanca,
